engine: Resources package rewrite
This giant patch makes some much needed improvements to the code base.

* The engine has been rewritten and lives within engine/graph/
* All of the common interfaces and code now live in engine/
* All of the resources are in one package called engine/resources/
* The Res API can use different "traits" from engine/traits/
* The Res API has been simplified to hide many of the old internals
* The Watch & Process loops were previously inverted, but is now fixed
* The likelihood of package cycles has been reduced drastically
* And much, much more...

Unfortunately, some code had to be temporarily removed. The remote code
had to be taken out, as did the prometheus code. We hope to have these
back in new forms as soon as possible.

+++ test/shell/env0.sh
@@ -9,12 +9,20 @@ set -o pipefail
 regex="123,,:123,321,:true,false:123"
 
 tmpdir="$($mktemp --tmpdir -d tmp.XXX)"
+if [[ ! "$tmpdir" =~ "/tmp" ]]; then
+	echo "unexpected tmpdir in: ${tmpdir}"
+	exit 99
+fi
 
 env TMPDIR="${tmpdir}" TEST=123 EMPTY="" $timeout -sKILL 60s "$MGMT" run --tmp-prefix --converged-timeout=5 --lang env0.mcl
 e=$?
 
 egrep "$regex" "$tmpdir/environ" || fail_test "Could not match '$(cat "$tmpdir/environ")' in '$tmpdir/environ' to '$regex'."
 
+if [ "$tmpdir" = "" ]; then
+	echo "BUG, tried to delete empty string path"
+	exit 99
+fi
 # cleanup if everything went well
 rm -r "$tmpdir"
