+++ test/test-crossbuild.sh
@@ -9,6 +9,10 @@ echo running "$(basename "$0")"
 # test if we can build for all OSes and ARCHes.
 
 tmpdir="`$mktemp --tmpdir -d tmp.XXX`"	# get a dir outside of the main package
+if [[ ! "$tmpdir" =~ "/tmp" ]]; then
+	echo "unexpected tmpdir in: ${tmpdir}"
+	exit 99
+fi
 log="$tmpdir/$(basename $0 .sh).log"
 
 set +e
@@ -20,6 +24,11 @@ if [ ! $RET -eq 0 ]; then
 	cat "$log"
 else
 	echo 'PASS'
+fi
+
+if [ "$tmpdir" = "" ]; then
+	echo "BUG, tried to delete empty string path"
+	exit 99
 fi
 rm -rf "$tmpdir"
 exit $RET
